requirements.txt
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
Deprecated==1.2.12
flake8==3.9.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mccabe==0.6.1
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
requests==2.26.0
six==1.16.0
urllib3==1.26.6
Werkzeug==2.0.1
wrapt==1.12.1
